#51 Reorganize routes
