+++ david/routes/api/__init__.py
@@ -1,0 +1,5 @@
+from sanic import Blueprint
+
+from .v1 import v1
+
+api = Blueprint.group([v1], url_prefix="/api")
