+++ david/routes/api/v1/__init__.py
@@ -1,0 +1,5 @@
+from sanic import Blueprint
+
+from .dialog import dialog
+
+v1 = Blueprint.group([dialog], url_prefix="/v1")
